Simplify scheduling of running mutations.
This just runs mutations as fast as they come in 🤷‍♂️ and rips
out all the complex timing stuff. The goal is to see how fast it
can go and sort of verify that what we are seeing is likely logical
bugs / maybe bad APIs and not perf limitations, as we as to create
a well performing baseline to build on top of.

Fixes #1